requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
tenacity>=8.2.0
selenium>=4.20.0
webdriver-manager>=4.0.0
